fix: round() 类型转换 - percent_rank 返回 double precision 需 ::numeric

+++ 05_代码与脚本/scripts/sql/biz/cm_onchain_percentile.sql
@@ -29,8 +29,8 @@ SELECT asset_id,
        metric,
        metric_date,
        value,
-       round(100 * pr_full, 2) AS pct_full,
-       round(100 * pr_roll, 2) AS pct_roll_365d,
+       round((100 * pr_full)::numeric, 2) AS pct_full,
+       round((100 * pr_roll)::numeric, 2) AS pct_roll_365d,
        CASE
          WHEN pr_full >= 0.90 THEN 'HIGH'
          WHEN pr_full <= 0.10 THEN 'LOW'
